requirements.txt
# Code Smell Detector Requirements

# Core dependencies
PyYAML>=6.0

# For testing
pytest>=7.4.0
pytest-cov>=4.1.0

